refactor: ビジネスロジックをルーティングから分離
- レポート生成ロジックを `server.js` から `report-generator.js` に分離。
- `server.js` はルーティングとデータ連携の責務に限定され、コードの見通しと保守性が向上。
